tempfix removed all ervice entries from db, need to change it in production

+++ node-manager/main.py
@@ -194,6 +194,7 @@ def initialize():
     module = module_data["modules"]
 
     for i, service in enumerate(module):
+        collection.delete_many({"name": service})
         cmd = f"docker stop {service} && docker rm {service}"
         os.system(cmd)
         cmd = f"docker rmi {service}"
@@ -213,6 +214,7 @@ def initialize():
 
 def create_node(service: str):
     collection = db.Services
+    collection.delete_many({"name": service})
     cmd = f"docker stop {service} && docker rm {service}"
     os.system(cmd)
     cmd = f"docker rmi {service}"
@@ -251,8 +253,8 @@ def start_node(service: str):
 def remove_node(service: str):
     collection = db.Service
     active = collection.find_one({"name": service})
-    # if not active:
-    #     return {"status": "False", "msg": "Node is not in our database, please create one"}
+    if not active:
+        return {"status": "False", "msg": "Node is not in our database, please create one"}
     cmd = f"docker stop {service} && docker rm {service}"
     os.system(cmd)
     cmd = f"docker rmi {service}"
